pandas, matplotlib, statistics
In this project we assume that we are working for an e-learning company. The company provides programming courses. Most of the curses are on web and mobile development, however we wll also cover many other domains (data science, game development etc.). Our company invests some money in advertisment in order to promote products.

The main goal of this project is to find out the two best markets to advertise the products in.